Flask_API/requirements.txt
Flask
psycopg2-binary
python-dotenv
PyJWT
